crewai/requirements.txt
pydantic
crewai
crewai-tools
langfuse
openinference-instrumentation-crewai
openinference-instrumentation-litellm
openlit
fastapi[standard]
dotenv